+++ tools/data.ts
@@ -1,0 +1,54 @@
+export const idolList = [
+  '櫻木真乃',
+  '風野灯織',
+  '八宮めぐる',
+  '月岡恋鐘',
+  '田中摩美々',
+  '白瀬咲耶',
+  '三峰結華',
+  '幽谷霧子',
+  '小宮果穂',
+  '園田智代子',
+  '西城樹里',
+  '杜野凛世',
+  '有栖川夏葉',
+  '大崎甘奈',
+  '大崎甜花',
+  '桑山千雪',
+  '芹沢あさひ',
+  '黛冬優子',
+  '和泉愛依',
+  '浅倉透',
+  '樋口円香',
+  '市川雛菜',
+  '福丸小糸',
+  '七草にちか',
+  '緋田美琴'
+]
+
+export const seriesList = [
+  {
+    name: 'ワンデイシリーズ',
+    regex: /ワンデイ(オフィサー|チーフ|キャプチャー)/
+  },
+  {
+    name: 'ワンダーランドシリーズ',
+    regex: /ワンダーランド(ホッピン|チェシャー|ハット|クイーン|アリス)/
+  },
+  {
+    name: 'パジャマ・デシリーズ',
+    regex: /パジャマ・デ・(カブ|ハム|ガオ|ペン|ワン){2}/
+  },
+  {
+    name: 'ティータイムシリーズ',
+    regex: /(キュー|テイス|ミン)ティータイムスイーツ/
+  },
+  {
+    name: 'メリーメリーシリーズ',
+    regex: /メリーメリー(デリバリー|ツリー|キャリー)/
+  },
+  {
+    name: 'クラシカルシリーズ',
+    regex: /クラシカル(メイド|マドモアゼル)/
+  }
+]
